requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
anthropic>=0.7.0
python-dotenv>=0.19.0
chromadb>=0.4.0
sentence-transformers>=2.0.0
pypdf2>=3.0.0
python-docx>=0.8.0
numpy>=1.21.0
vercel>=0.0.0
